Properly handle malformed request payload.
If the request contains malformed payload (i.e. no proper JSON), then
the creation of the EventBusMessage fails and trying to create a
response from it fails with a runtime exception that is swallowed
without sending back a response to the client.

The error response message is now always created from the original
AMQP request message only.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>