Fix Spring Boot configuration.
Spring Boot's mapping mapping of configuration properties based on a
prefix requires the configuration properties to use exactly the prefix
in their name. The "relaxed binding" is only used for mapping to the
Java class properties.

In order to have a consistent approach for the property names in our
YAML config files, the Device Connection config properties now use the
"hono.devcon" prefix which does not require us to introduce kebab case
property names to the YAML files.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>