[#1302] Configure documentation of 1.0-M5 as current "stable" version.
Changed links that are supposed to point to the stable version back to
"stable" (as planned with issue #1302) and set the latest milestone
(1.0-M5) as the stable version. For the transition phase until the 1.0.0
release, the latest milestone is intended to be marked as "stable" as far
as the documentation is concerned.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>