[#1302] Set 1.0-M5 docu as "stable" and improve website pipeline.
Change links that are supposed to point to the stable version back to
"stable" (as planned with issue #1302) and set the latest milestone
(1.0-M5) as the stable version. For the transition phase until the 1.0.0
release, the latest milestone is intended to be marked as "stable" as
far as the documentation is concerned.
The homepage directory is now copied at the beginning of the jenkins
pipeline to make sure that is always build from master and to prevent
conflicts with changed files when switching branches.
Improve calculation of the order of versions in the menu.
Make sure that the "old version hint" banner always points the current
stable version.
Display "Getting started" as first menu item on homepage.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>