[#1414] Remove dependency on Jaeger Operator.
The chart doesn't depend on the Jaeger Operator anymore. Instead, the
standard Jaeger all-in-one image is used to deploy an example Jaeger
back end. The chart also supports using an existing Jaeger back end by
means of setting configuration variables on the Jaeger Agent that is run
as a sidecar container with all Hono components.

Fixes #1414

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>